Remove all old testing code from the test library
