--- orbit-test/src/iosMain/kotlin/org/orbitmvi/orbit/Blocking.kt
@@ -1,23 +1,0 @@
-/*
- * Copyright 2021 Mikołaj Leszczyński & Appmattus Limited
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.orbitmvi.orbit
-
-import kotlin.coroutines.CoroutineContext
-
-internal actual fun <T> runBlocking(block: suspend () -> T): T = kotlinx.coroutines.runBlocking { block() }
-internal actual fun <T> runBlocking(coroutineContext: CoroutineContext, block: suspend () -> T): T =
-    kotlinx.coroutines.runBlocking(coroutineContext) { block() }
